First commit of the core parts written over the past few days to get started. Most of the back end is in place and organized for potential reuse.
The front end has not yet been implemented so there is no content assist stype popup implemented yet, no plugin preferences, and the class written for pulling the templates from a Confluence WIKI has not been tied in.

Currently it does parse form definitions from the first found comment section of the template file (*.ftl) and builds the form model from that XML. Render methods have been written for that form model using SWT to generate a form for each prompt group with each prompt included as an input element on the form. Supported input types are: text, multiline text, date, and checkbox.

A form data object has been designed for storing the resulting data from the dynamically rendered forms. A merge method written for combining that form data with the template to produce the resulting document into a string. The code has not yet been implemented to merge the resulting string with the active source member in the LPEX editor.